Move verification state and errors into SDK
